telegram_bot/requirements.txt
python-telegram-bot==21.0.1
APScheduler==3.10.4
python-dotenv==1.0.1
openai==1.14.3
